+++ src/docs/components/callout.svelte
@@ -2,7 +2,7 @@
 	import { tv, type VariantProps } from 'tailwind-variants';
 	import { cn } from '$docs/utils';
 	const calloutVariants = tv({
-		base: 'relative rounded-tr-md rounded-br-md px-5 py-3 before:absolute before:left-0 before:top-0 before:h-full before:w-0.5 before:content-[""] my-6 text-sm',
+		base: 'relative rounded-tr-md rounded-br-md px-5 py-3 before:absolute before:left-0 before:top-0 before:h-full before:w-0.5 before:content-[""] my-6 ',
 		variants: {
 			type: {
 				default: 'before:bg-magnum-700 bg-magnum-700/10 text-white border-magnum-700',
@@ -19,8 +19,6 @@
 	export let type: VariantProps<typeof calloutVariants>['type'] = 'default';
 </script>
 
-<div class={cn(calloutVariants({ type, className }))}>
-	<p>
-		<slot />
-	</p>
+<div class={cn(calloutVariants({ type, className }))} data-callout>
+	<slot />
 </div>