+++ src/docs/components/type-dialog.svelte
@@ -41,7 +41,8 @@
 			</div>
 			<button
 				melt={$close}
-				class="absolute right-[10px] top-[10px] inline-flex h-[25px] w-[25px]
+				aria-label="Close"
+				class="absolute right-[10px] top-[10px] inline-flex h-6 w-6
                 appearance-none items-center justify-center rounded-full text-magnum-300
                 hover:bg-magnum-800/50 focus:shadow-magnum-400"
 			>