+++ src/lib/builders/accordion/create.ts
@@ -8,21 +8,28 @@ import {
 	kbd,
 } from '$lib/internal/helpers';
 import type { Defaults } from '$lib/internal/types';
-import { derived, get, writable } from 'svelte/store';
-import type { AccordionItemProps, CreateAccordionProps } from './types';
+import { tick } from 'svelte';
+import { derived, writable, type Writable } from 'svelte/store';
+import type {
+	AccordionHeadingProps,
+	AccordionItemProps,
+	AccordionType,
+	CreateAccordionProps,
+} from './types';
 
-type AccordionParts = 'trigger' | 'item' | 'content';
+type AccordionParts = 'trigger' | 'item' | 'content' | 'heading';
 const { name, selector } = createElHelpers<AccordionParts>('accordion');
 
 const defaults = {
 	type: 'single',
 } satisfies Defaults<CreateAccordionProps>;
 
-export const createAccordion = (props?: CreateAccordionProps) => {
-	const withDefaults = { ...defaults, ...props } as CreateAccordionProps;
+export const createAccordion = <T extends AccordionType = 'single'>(
+	props?: CreateAccordionProps<T>
+) => {
+	const withDefaults = { ...defaults, ...props } as CreateAccordionProps<T>;
 	const options = writable({
 		disabled: withDefaults.disabled,
-		type: withDefaults.type,
 	});
 
 	const value = writable<string | string[] | undefined>(withDefaults.value);
@@ -55,6 +62,14 @@ export const createAccordion = (props?: CreateAccordionProps) => {
 		}
 	};
 
+	const parseHeadingProps = (props: AccordionHeadingProps) => {
+		if (typeof props === 'number') {
+			return { level: props };
+		} else {
+			return props;
+		}
+	};
+
 	const item = builder(name('item'), {
 		stores: value,
 		returned: ($value) => {
@@ -74,26 +89,26 @@ export const createAccordion = (props?: CreateAccordionProps) => {
 		returned: ([$value, $options]) => {
 			return (props: AccordionItemProps) => {
 				const { value: itemValue, disabled } = parseItemProps(props);
-
+				// generate the content ID here so that we can grab it in the content
+				// builder action to ensure the values match.
 				return {
-					'aria-expanded': isSelected(itemValue, $value) ? true : false,
 					disabled: $options.disabled || disabled,
+					'aria-expanded': isSelected(itemValue, $value) ? true : false,
+					'aria-disabled': disabled ? true : false,
 					'data-disabled': disabled ? true : undefined,
 					'data-value': itemValue,
-					// TODO: aria-controls, aria-labelledby
 				};
 			};
 		},
 		action: (node: HTMLElement) => {
 			const unsub = executeCallbacks(
 				addEventListener(node, 'click', () => {
-					const $options = get(options);
 					const disabled = node.dataset.disabled === 'true';
 					const itemValue = node.dataset.value;
 					if (disabled || !itemValue) return;
 
 					value.update(($value) => {
-						if ($options.type === 'single') {
+						if (withDefaults.type === 'single') {
 							return $value === itemValue ? undefined : itemValue;
 						} else {
 							const arrValue = $value as string[] | undefined;
@@ -116,22 +131,23 @@ export const createAccordion = (props?: CreateAccordionProps) => {
 					const el = e.target as HTMLElement;
 					const rootEl = getElementByMeltId(ids.root);
 					if (!rootEl) return;
-					const items = Array.from(rootEl.querySelectorAll(selector('trigger'))) as HTMLElement[];
+					const items = Array.from(rootEl.querySelectorAll<HTMLElement>(selector('trigger')));
+					const candidateItems = items.filter((item) => item.dataset.disabled !== 'true');
 
-					if (!items.length) return;
-					const elIdx = items.indexOf(el);
+					if (!candidateItems.length) return;
+					const elIdx = candidateItems.indexOf(el);
 
 					if (e.key === kbd.ARROW_DOWN) {
-						items[(elIdx + 1) % items.length].focus();
+						candidateItems[(elIdx + 1) % candidateItems.length].focus();
 					}
 					if (e.key === kbd.ARROW_UP) {
-						items[(elIdx - 1 + items.length) % items.length].focus();
+						candidateItems[(elIdx - 1 + candidateItems.length) % candidateItems.length].focus();
 					}
 					if (e.key === kbd.HOME) {
-						items[0].focus();
+						candidateItems[0].focus();
 					}
 					if (e.key === kbd.END) {
-						items[items.length - 1].focus();
+						candidateItems[candidateItems.length - 1].focus();
 					}
 				})
 			);
@@ -151,7 +167,36 @@ export const createAccordion = (props?: CreateAccordionProps) => {
 				return {
 					'data-state': selected ? 'open' : 'closed',
 					'data-disabled': $options.disabled ? true : undefined,
+					'data-value': itemValue,
 					hidden: selected ? undefined : true,
+				};
+			};
+		},
+		action: (node: HTMLElement) => {
+			tick().then(() => {
+				const contentId = generateId();
+				const triggerId = generateId();
+
+				const parentTrigger = document.querySelector<HTMLElement>(
+					`${selector('trigger')}, [data-value="${node.dataset.value}"]`
+				);
+				if (!parentTrigger) return;
+
+				node.id = contentId;
+				parentTrigger.setAttribute('aria-controls', contentId);
+				parentTrigger.id = triggerId;
+			});
+		},
+	});
+
+	const heading = builder(name('heading'), {
+		returned: () => {
+			return (props: AccordionHeadingProps) => {
+				const { level } = parseHeadingProps(props);
+				return {
+					role: 'heading',
+					'aria-level': level,
+					'data-heading-level': level,
 				};
 			};
 		},
@@ -159,11 +204,13 @@ export const createAccordion = (props?: CreateAccordionProps) => {
 
 	return {
 		root,
-		value,
+		/** Initial value of accordion */
+		value: value as Writable<CreateAccordionProps<T>['value']>,
 		item,
 		trigger,
 		content,
 		isSelected: isSelectedStore,
 		options,
+		heading,
 	};
 };