+++ src/lib/builders/tooltip/create.ts
@@ -88,7 +88,9 @@ export function createTooltip(props?: CreateTooltipProps) {
 				addEventListener(node, 'mouseout', () => get(closeTooltip)()),
 				addEventListener(node, 'focus', () => open.set(true)),
 				addEventListener(node, 'blur', () => open.set(false)),
-				addEventListener(node, 'pointerdown', () => {
+				addEventListener(node, 'mousedown', (e) => {
+					e.preventDefault();
+
 					const $options = get(options);
 					if ($options.closeOnPointerDown) {
 						open.set(false);